services/insights-llm/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
mysql-connector-python